feat(protocol): widen NetHeader::origin_hash to u64
WIRE_ORIGIN_HASH_64BIT cutover, steps 1-6 of the plan. Wire format
break — old + new nodes cannot interoperate at the packet layer.

protocol.rs:
- HEADER_SIZE: 64 → 68 (wire size).
- #[repr(C, align(64))] → #[repr(C, align(8))] on NetHeader; the
  prior cache-line alignment was reclaimed when the size left the
  cache-line boundary.
- origin_hash field widens u32 → u64.
- Fields reordered: origin_hash precedes subnet_id so the u64 sits
  at the natural 8-aligned offset 48; otherwise the in-memory
  layout would force 4 bytes of padding before origin_hash.
- AAD grows 52 → 56 bytes (origin_hash widens; subnet_id shifts).
- to_bytes / from_bytes flip put_u32_le/get_u32_le → u64 variants.
- with_origin drops the `as u32` truncation cast.
- Layout diagram updates: ORIGIN_HASH now 8 bytes ending at 56;
  SUBNET_ID at 60; FRAGMENT_ID/OFFSET at 64; PAYLOAD_LEN/EVENT_COUNT
  at 68.
- size_of::<NetHeader>() is 72 (align(8) rounds up); HEADER_SIZE is
  the 68-byte wire size. The static + runtime assertions split:
  one pins the wire-size constant, the other pins the in-memory
  size against regression.

mesh.rs:
- origin_hash_to_node populator (failure-detector eviction +
  announcement handler) drops the `as u32 as u64` truncation; keys
  on the full EntityId::origin_hash() u64.

cortex/rpc.rs:
- RpcInboundEvent::origin_hash widens u32 → u64 to mirror the new
  wire field.

Step 7 (OriginHashSlot::Multiple → Option<NodeId> collapse) and
tests follow in subsequent commits.

+++ net/crates/net/src/adapter/net/mesh.rs
@@ -1947,10 +1947,11 @@ impl MeshNode {
             // releases its claim on the wire hash.
             let removed_entity_id = peer_entity_ids_failure.remove(&node_id).map(|(_, eid)| eid);
             if let Some(eid) = removed_entity_id {
-                // Same wire-truncated u32-as-u64 key the inbound
-                // announcement handler uses; see the comment above
-                // `ctx.origin_hash_to_node.entry(origin_hash)`.
-                let origin_hash = eid.origin_hash() as u32 as u64;
+                // Same key the inbound announcement handler uses —
+                // the full 64-bit `EntityId::origin_hash()` since the
+                // `WIRE_ORIGIN_HASH_64BIT` cutover (wire header
+                // now carries the full u64 too).
+                let origin_hash = eid.origin_hash();
                 let mut drop_slot = false;
                 if let Some(mut slot) = origin_hash_to_node_failure.get_mut(&origin_hash) {
                     drop_slot = slot.remove(node_id);
@@ -5827,14 +5828,11 @@ impl MeshNode {
                 // node_id; a distinct node_id claiming the same
                 // wire hash promotes the slot to Multiple, which
                 // future get_node_by_origin_hash calls resolve
-                // to None.
-                // Wire packet headers carry `origin_hash` as a u32
-                // (see `protocol::PacketHeader::origin_hash`); the
-                // channel-publish path stamps `entity_id.origin_hash() as u32`.
-                // Key the reverse index on the SAME truncated form so
-                // greedy-admission lookups against `parsed.header.origin_hash`
-                // hit the slot we just installed.
-                let origin_hash = ann.entity_id.origin_hash() as u32 as u64;
+                // to None. Post-`WIRE_ORIGIN_HASH_64BIT` the wire
+                // header carries the full `EntityId::origin_hash()`
+                // u64, so the reverse-index key matches the
+                // application-layer value verbatim.
+                let origin_hash = ann.entity_id.origin_hash();
                 ctx.origin_hash_to_node
                     .entry(origin_hash)
                     .and_modify(|slot| {
